Add Safari iOS versions for api.TouchEvent.altKey (#13843)
This PR adds real values for Safari iOS/iPadOS for the `altKey` member of the `TouchEvent` API by mirroring the data.  Since this feature was available in the first version of Chrome this API was introduced in, this property would most likely be around in the first Safari iOS version of the API as well.

Note: even though iOS didn't have the same level of keyboard support back then as it does now, the properties were still around.  This matches the data for the other keyboard properties (Ctrl, meta, shift).