Fix Chromium data for devicemotion event interfaces (#14445)
This was all added in WebKit trunk 534.5 and 534.9:
https://trac.webkit.org/changeset/64845/webkit
https://trac.webkit.org/changeset/68252/webkit

It wasn't enabled then, but because it all went together, it can all be
assumed to match the ondevicemotion property being exposed. That was
confirmed to be Chrome 31 with this test:
http://mdn-bcd-collector.appspot.com/tests/api/Window/ondevicemotion

The source of Chrome 11 is here:
https://github.com/mdn/browser-compat-data/pull/5834

The description says Chrome 31, so this must have been a copy-paste
error from the IE version.

Also update the Samsung Internet and WebView versions based on
conservative mirroring. It's possible this was supported earlier, but
the current version numbers also seem to be based on mirroring, not
testing:
https://github.com/mdn/browser-compat-data/pull/6838